data_collector/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
mysql-connector-python==8.0.15
yfinance==0.2.50
